// frontend/src/models/PairDetails.ts
type PairDetails = {
  token1Name: string;
  token1Symbol: string;
  token1Index: number;
  token2Name: string;
  token2Symbol: string;
  token2Index: number;
  liquidityTokenName: string;
  liquidityTokenSymbol: string;
  liquidityTokenIndex: number;
  escrowAddress: string;
  escrowLogicSig: string;
};

export default PairDetails;